# GO_Backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Social Media for musicians — Backend API",
        "title": "LickLib API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/auth/login": {
            "post": {
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Username",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Passwort",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/tracks": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "tracks"
                ],
                "summary": "Track hochladen",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Audiodatei (MP3/WAV/FLAC, max 100MB)",
                        "name": "trackFile",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Titel (3-200 Zeichen)",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Beschreibung (10-2000 Zeichen)",
                        "name": "description",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "EASY|MEDIUM|HARD|GOGGINS",
                        "name": "difficulty",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/tracks/user/{username}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tracks"
                ],
                "summary": "Alle Tracks eines Users abrufen",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Username",
                        "name": "username",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Track"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/tracks/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tracks"
                ],
                "summary": "Track per ID abrufen",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Track UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Track"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "tracks"
                ],
                "summary": "Track löschen",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Track UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tracks"
                ],
                "summary": "Track-Metadaten updaten",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Track UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update-Felder",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.UpdateTrackRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/tracks/{id}/play": {
            "get": {
                "tags": [
                    "tracks"
                ],
                "summary": "Track abspielen (Presigned URL)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Track UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "307": {
                        "description": "Temporary Redirect"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.Difficulty": {
            "type": "string",
            "enum": [
                "EASY",
                "MEDIUM",
                "HARD",
                "GOGGINS"
            ],
            "x-enum-varnames": [
                "DifficultyEasy",
                "DifficultyMedium",
                "DifficultyHard",
                "DifficultyGoggins"
            ]
        },
        "models.Notation": {
            "type": "object",
            "properties": {
                "author": {
                    "$ref": "#/definitions/models.User"
                },
                "authorID": {
                    "type": "integer"
                },
                "content": {
                    "description": "Tab oder Noten als Text/JSON",
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "fileExt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "sizeBytes": {
                    "type": "integer"
                },
                "track": {
                    "$ref": "#/definitions/models.Track"
                },
                "trackID": {
                    "type": "integer"
                },
                "type": {
                    "description": "\"TABS\" | \"NOTES\"",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.NotationType"
                        }
                    ]
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.NotationType": {
            "type": "string",
            "enum": [
                "TABS",
                "NOTES"
            ],
            "x-enum-varnames": [
                "NotationTypeTabs",
                "NotationTypeNotes"
            ]
        },
        "models.Track": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "difficulty": {
                    "$ref": "#/definitions/models.Difficulty"
                },
                "file_ext": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "notations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Notation"
                    }
                },
                "size_bytes": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/models.TrackStatus"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/models.User"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.TrackStatus": {
            "type": "string",
            "enum": [
                "UPLOADING",
                "READY",
                "FAILED",
                "PROCESSING"
            ],
            "x-enum-varnames": [
                "TrackStatusUploading",
                "TrackStatusReady",
                "TrackStatusFailed",
                "TrackStatusProcessing"
            ]
        },
        "models.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "description": "PasswordHash string     `gorm:\"column:password_hash;type:text;not null\" json:\"-\"`",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "notations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Notation"
                    }
                },
                "tracks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Track"
                    }
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "service.UpdateTrackRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}